// src/module/types/types.ts
export enum EstadoAvaliacaoEnum {
  FREQUENCIA_2 = '2ª Frequência',
  EXAME = 'Exame',
  RECURSO = 'Recurso',
  APROVADO = 'Aprovado',
  REPROVADO = 'Reprovado',
  AGUARDA_PRATICA = 'Aguarda Nota da Prática',
  AGUARDA_ORAL = 'Aguarda Nota da Oral',
  AGUARDA_ORAL_RECURSO = 'Aguarda Nota da Oral do Recurso',
  AGUARDA_ORAL_EXAME_ESPECIAL = 'Aguarda Nota da Oral Exame Especial',
  PENDENTE = 'Sem Avaliações',
  AGUARDA_RESULTADO = 'Aguarde Resultado',
}